Rename "error" module to "error_old"
I'm going to make a new one, let's rename this one so they can co-exist.
