+++ src/parser.rs
@@ -8,7 +8,7 @@
 //! The parser converts a sequence of tokens into a Concrete Syntax Tree.
 
 use crate::cst::{BinOp, Expr, NonCode, Prefixed, Seq, Stmt, StringPart, UnOp};
-use crate::error::{IntoParseError, ParseError};
+use crate::error_old::{IntoParseError, ParseError};
 use crate::lexer::{Lexeme, QuoteStyle, StringPrefix, Token};
 use crate::source::{DocId, Span};
